ednftm_v2/requirements.txt
torch>=2.0.0
numpy>=1.24.0
scipy>=1.12.0
scikit-learn>=1.3.0
gensim==4.3.2
nltk>=3.8.0
pyyaml>=6.0
tqdm>=4.65.0
sentence-transformers>=2.2.0
transformers>=4.30.0
